test_main: avoid signals for timeout handling
Previously we used setitimer() to trigger test timeouts after the
appropriate delay. However, this caused the timeout code to run in a
signal handler, which TSAN flagged as unsafe because it allocates
memory.

This changes it to just use its own thread which sleeps before aborting
the process after the prescribed timeout.

The resulting behavior should be the same, though will avoid a bunch of
confusing spew in TSAN builds that time out.

Change-Id: I0af2485c14db366a409ba1409fa01314396d030a
Reviewed-on: http://gerrit.cloudera.org:8080/4415
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>