KUDU-1596. Implement a version migration test
This program takes two binary directories and runs a simple "migration"
test between the two versions to ensure that a table created with
"Version A" of the software can be read by "Version B" of the software.
This can also be thought of as an "upgrade" or "downgrade" test.

This patch also makes some changes to ExternalMiniCluster and the cmake
files to allow for creating an executable that is written like a unit
test (using the Google Test framework) but is not automatically executed
by ctest.

As a follow-up, we would benefit from a higher-level script that can
manage upgrade and downgrade testing across many versions of Kudu.

Tested with Version A = 0.10.0 and Version B = 1.0.0

Change-Id: I2831b47e6c0b644a256e1914fa495f453318e0cd
Reviewed-on: http://gerrit.cloudera.org:8080/4424
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>