[c++client] added TestAutoFlushBackgroundDropSession
As a follow-up for 1a062253e3fdc900a4b0b418520d2870b6de8846,
added new TestAutoFlushBackgroundDropSession test.
The test verifies that a session in AUTO_FLUSH_BACKGROUND mode can
be safely abandoned so its pending operations are not flushed.

Also, reduced number of write operations applied to the session and
updated the limit on mutation buffer size accordingly in
TestApplyTooMuchWithoutFlushing test to reduce running time for
ASAN/TSAN configurations.  The semantics of the test is to try
adding more data into the mutation buffer than the limit on the buffer
size: the crux here is in the relative sizes, not the absolute ones.

Change-Id: If136872a4200e18acbb587276fed28975afe3810
Reviewed-on: http://gerrit.cloudera.org:8080/4432
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>