make election timeout jitter more aggressive
Random election timeout jitter is necessary in Raft in order to
guarantee that an election can be won. If the jitter is smaller than RTT
or the accuracy of clocks, then elections could fail indefinitely. We
frequently hit an issue during tests where timeouts tend to 'clump'
together, causing elections to retry many times in a row, ultimately
leading to test timeout. This commit increases the jitter, so that
election timeout differences between nodes will hopefully be greater
than the clock error. This issue could also manifest if the RTT between
nodes is high.

Change-Id: I2c9dad820c2b7d4bc4b9e791b78222559cdf63c8
Reviewed-on: http://gerrit.cloudera.org:8080/3828
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>