Add support for clang from XCode 8 in macOS
Since upgrading XCode to 8 (even in El Capitan), cmake has outputted an
unsupported compiler message:

  Unknown compiler.  Version info:

  Apple LLVM version 8.0.0 (clang-800.0.38)

  Target: x86_64-apple-darwin16.0.0

  Thread model: posix

  InstalledDir:
  /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin

This is also the default version in the new macOS Sierra.
This patch fixes this by adding it to the list of supported
compilers.

Note that this patch, by itself, won't allow to fully compile Kudu in
macOS Sierra, as the current version of tcmalloc is incompatible:
https://github.com/gperftools/gperftools/issues/827

We need to wait until the issue above is fixed in order to fully support
building with the default toolchain in macOS Sierra. In the meanwhile this
patch is still valuable. It's still possible to build Kudu with the clang
toolchain we ship and CLion users can still use the default toolchain,
as long as they don't actually run the CLion build's binaries.

Were we to change to jemalloc, that issue is solved there:
https://github.com/jemalloc/jemalloc/pull/427

Change-Id: I74574fb80525c7b7545e2aaf9b220f63486a46b6
Reviewed-on: http://gerrit.cloudera.org:8080/4482
Reviewed-by: Dan Burkert <dan@cloudera.com>
Tested-by: Kudu Jenkins