tsan-suppressions: suppress various glog/gflags data races
In an upcoming patch we'll begin building more of our thirdparty
dependencies with -fsanitize=thread. Having tested that, I saw that TSAN now
flags some new data races. All of them are "benign" in that they should be
safe on x86. Nonetheless, we either need to patch glog/gflags to fix them,
or suppress them. I chose the latter for expediency.

Change-Id: I3906070cb0e202e5b841f5b163e4adc023fe3882
Reviewed-on: http://gerrit.cloudera.org:8080/4509
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>