c++ client: adjust kudu::client::sp
In an upcoming change we'll begin using libc++ on Linux for TSAN builds. As
macOS developers are probably aware, libc++ lacks TR1 support, so we need to
extend our existing macOS-specific workaround for TR1 shared_ptrs to apply
to any libc++ build.

This has no bearing on backwards compatibility as libc++ is still only used
in non-shipping environments.

Change-Id: I01a1ef6319c3464a8ec84f066adc885da200af70
Reviewed-on: http://gerrit.cloudera.org:8080/4510
Reviewed-by: Dan Burkert <dan@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>