Patch gperftools to be able to work on macOS Sierra/XCode 8
Apparently gperftools needs to be patched for tcmalloc to work
with the new clang version that ships with XCode 8, but the fix
hasn't landed yet.

This patch fixes that and was copied verbatim from:
https://github.com/gperftools/gperftools/issues/827

Note that this fix is virtually identical to the one merged into
jemalloc, which had the same problem:
https://github.com/jemalloc/jemalloc/pull/427/commits/19c9a3e828ed46f1576521c264640e60bd0cb01f

Finally it's noteworthy that this is an macOS/OS X only
fix and, as such, doesn't affect released binaries.

Change-Id: Ie00d9fad36d1746aa94da3b1711353110da42d95
Reviewed-on: http://gerrit.cloudera.org:8080/4495
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins