KUDU-1651: Check empty dict before decoder eval
A completely null dictionary-encoded column will yield an empty
dictionary. When decoder-level evaluation is on, this will result in
an error when creating the bitmap of matching codewords, as bitmaps
must have a non-zero size.

In this scenario, it is sufficient to not determine the set of
matching codewords. The decoder should be completely skipped over
given a null column, so the matching codeword set would not be used
here. If the decoder is not skipped over and an attempt is made to
read from the dictionary block, an error should be thrown.

Change-Id: I76d457c65238b3b2283d826076fdc7525d032e3f
Reviewed-on: http://gerrit.cloudera.org:8080/4537
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>
(cherry picked from commit 1f40913de6d3427847f3d435e1b84fb928a6f6a9)
Reviewed-on: http://gerrit.cloudera.org:8080/4606
Reviewed-by: Todd Lipcon <todd@apache.org>