KUDU-1654 - [python] Python 3 Client Test Failure: test_table_column
Python 3 fails the test_table_column test because it is is attempting
to cast a string to bytes without an encoding argument. In Python 3,
this is a required parameter as bytes and strings are no longer
synonymous. This patch utilizes the kudu.compat.frombytes method
which was written for this purpose.

Change-Id: I7184639426b7f2528523209152dafd5fa39fecf9
Reviewed-on: http://gerrit.cloudera.org:8080/4543
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>