thirdparty: clean up llvm cmake files too
The upgrade to LLVM 3.9 changed the list of targets provided in LLVM's cmake
files. If the cmake files are left behind and make it into an LLVM 3.8-based
build, the subsequent Kudu build will fail because it can't find all of the
targets listed.

It's too late to help with the LLVM 3.9 upgrade, but hopefully it'll ease
future upgrades.

Change-Id: I98393cf1f8afc2ced78245cad5e2e24ee9410214
Reviewed-on: http://gerrit.cloudera.org:8080/4549
Reviewed-by: Dan Burkert <dan@cloudera.com>
Tested-by: Kudu Jenkins