rpc: Add min / max negotiation threads
Since we use a thread pool for connection negotiation, it would be
helpful for users to be able to specify the minimum and maximum number
of threads in that pool. Prior to this patch, there were no gflags to
control these parameters, and the only builder parameter was the maximum
(however it was not labelled as such).

This allows tuning of the per-tserver negotiation pool size, both
minimum and maximum threads to allow in the thread pool.

Change-Id: Ife98b39d5f3a340702151ab27dc8026c8bac12ac
Reviewed-on: http://gerrit.cloudera.org:8080/4574
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>