Retry on early TimedOut error in ExternalMiniCluster::WaitForCatalogManager
WaitForCatalogManager loops sending a GetTableLocations RPC to a master
waiting for it to be initialized. Previously, we would early return with
an error when the RPC resulted in a TimedOut response, despite the
method having its own time out limit. This was causing at least some of
the flakiness seen in alter_table-randomized-itest.

Change-Id: I04b3a58646249a0fd549c75dc130ca9b955f35de
Reviewed-on: http://gerrit.cloudera.org:8080/4576
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins