KUDU-1652 (part 2): Filter IS NOT NULL predicates from scan on client
Part 1 fixed the C++ client/server and Java client so that specifying a
an IS NOT NULL predicate on a non-nullable column would not result in a
CHECK failure or exception. This commit filters all such predicates from
new scan requests on the client, in order to avoid crashing server
versions without the part 1 fix (1.0.0 and below).

This is a difficult thing to test, since the underlying issue has
already been fixed, and this just prevents older server version from
crashing. I've manually verified the change by testing this patch with
the additional tests in part 1, but without the fix from part 1.

Change-Id: I542a735ca337ec9ed4eb3c989f8e6fa4ee0cc1da
Reviewed-on: http://gerrit.cloudera.org:8080/4612
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Dan Burkert <dan@cloudera.com>