raft_consensus: clean up overrides, protected methods
RaftConsensus used to have a subclass mock in raft_consensus-test, but
that test is now removed. Given this, we can make some
previously-protected methods private and non-virtual.

I also took this opportunity to modernize the 'override' style to C++11,
and also to remove a function which was essentially duplicate code.

Change-Id: Ib52d3ac40ed68a4ff8b1738897f6ef62f94a843b
Reviewed-on: http://gerrit.cloudera.org:8080/4620
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>